refactor(32): corrects incorrect casing in file name
