+++ src/mock/entities/VehicleType.ts
